feat (store/SaleDone): add comission to affilate when sale its done

+++ affiliates/models.py
@@ -29,7 +29,7 @@ class Affiliate(models.Model):
         null=True,
         blank=True,
     )
-    balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
+    balance = models.FloatField(default=0.00)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
